Turn the remaining 'sh' snippets into console ones
Mostly because these are not snippets in the bash programming language
but they show you what commands to copy&paste.

Also remove the leading dollar sign, unless the snippet shows both a
typed command its output. This helps when tripple-click is used to
select the full line and the dollar sign would get into the reader's
way.