create, read, update: handle created & modified property for passwords
The modified flag is handy when you want to make sure the password on a
given service is new enough (passwords before a certain time are
considered unsafe). And if we are at it, also add a creation time, too.

Store the date/time in a format like 2024-10-12T20:03:46+02:00, which is
machine-readable, but still more (human-)readable than just git-style
epoch seconds + timezone info.